+++ CHANGELOG.md
@@ -4,5 +4,6 @@ CHANGELOG
 dev-master
 ----------
 
+- XML driver handles meta fields Url, Title and Description
 - Fire event for each found search result hit to allow listeners to modify the document (e.g.
   set URL or other data on the document).